+++ apps/client/src/app/services/impersonation-storage.service.ts
@@ -1,11 +1,9 @@
-import { Injectable } from '@angular/core';
+import { Service } from '@angular/core';
 import { BehaviorSubject } from 'rxjs';
 
 export const IMPERSONATION_KEY = 'impersonationId';
 
-@Injectable({
-  providedIn: 'root'
-})
+@Service()
 export class ImpersonationStorageService {
   private hasImpersonationChangeSubject = new BehaviorSubject<string | null>(
     this.getId()